Adding edit functionality

+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {MatSort} from '@angular/material/sort';
 import {MatTableDataSource} from '@angular/material/table';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {FormBuilder, FormGroup, FormGroupDirective, Validators} from '@angular/forms';
 
 import {Expense} from './expense.model';
 import {AngularFirestore} from '@angular/fire/firestore';
@@ -19,13 +19,14 @@ export class AppComponent implements OnInit{
   title = 'expense-manager';
   cards: Card[] = ['Forex', 'TranseferWise', 'CreditSuisse', 'UbsDebit', 'UbsCredit'];
   users: User[] = ['Ansuman', 'Debasrita'];
-  displayedColumns: string[] = ['user', 'card', 'date', 'amount', 'id'];
+  displayedColumns: string[] = ['user', 'card', 'date', 'amount', 'where', 'id'];
   expenses: Expense[] = [];
   dataSource: MatTableDataSource<Expense>;
   expenseForm: FormGroup;
   totalExpense: number;
   @ViewChild(MatSort, {static: true}) sort: MatSort;
   @ViewChild('toast', {static: true}) toast;
+  @ViewChild('formDirective', {static: true}) formDirective: FormGroupDirective;
 
   constructor(private fb: FormBuilder, private firestore: AngularFirestore, private snackBar: MatSnackBar) {
 
@@ -47,30 +48,43 @@ export class AppComponent implements OnInit{
 
   private createExpenseForm() {
     this.expenseForm = this.fb.group({
+      id: [''],
       user: ['', [Validators.required]],
       amount: ['', [Validators.required]],
       date: ['', [Validators.required]],
-      card: ['', [Validators.required]]
+      card: ['', [Validators.required]],
+      where: [''],
     });
 
     this.setDefaultExpense();
   }
 
+  onSubmit() {
+    const formData = Object.assign({}, this.expenseForm.value);
+    delete formData.id;
+    if (this.expenseForm.value.id == null) {
+      this.firestore.collection('expense').add(formData);
+    } else {
+      this.firestore.doc('expense/' + this.expenseForm.value.id).update(formData);
+    }
+    this.snackBar.openFromTemplate(this.toast, {
+      duration: 1000,
+    });
+    this.createExpenseTable();
+    this.resetForm();
+  }
+
   setDefaultExpense() {
-    this.expenseForm.setValue({
+    this.expenseForm.patchValue({
+      id: null,
       user: 'Ansuman',
-      amount: null,
       date: new Date(),
       card: 'TranseferWise',
     });
   }
 
-  onSubmit() {
-    this.firestore.collection('expense').add(this.expenseForm.value);
-    this.snackBar.openFromTemplate(this.toast, {
-      duration: 1000,
-    });
-    this.createExpenseTable();
+  resetForm() {
+    this.formDirective.resetForm();
     this.setDefaultExpense();
   }
 
@@ -80,12 +94,19 @@ export class AppComponent implements OnInit{
   }
 
   deleteExpense(id: string) {
-    this.firestore.doc<Expense>(`expense/${id}`).delete().then( _ => {
-      this.snackBar.open('Expense deleted');
-      setTimeout( () => {
-        this.snackBar.dismiss();
-      }, 2000);
-    });
+    const conf = confirm('Are you sure to delete?');
+    if (conf) {
+      this.firestore.doc<Expense>(`expense/${id}`).delete().then( _ => {
+        this.snackBar.open('Expense deleted');
+        setTimeout( () => {
+          this.snackBar.dismiss();
+        }, 2000);
+      });
+    }
   }
 
+  editExpense(expense: any) {
+    this.expenseForm.setValue(expense);
+    this.expenseForm.get('date').setValue(new Date(expense.date.seconds * 1000));
+  }
 }
